BEEFY: `expect_validator_set()` fix (#2716)
Fixes #https://github.com/paritytech/polkadot-sdk/issues/2699

Modifying `expect_validator_set()` in order to be able to walk back
until block 0. The chain state at block 0 is available even if we use
`--sync fast` or `--sync warp`. This way we can retrieve the initial
authority set even when BEEFY genesis is 1 and there is no authority
change entry in the headers log.

Credits to @acatangiu for the solution

---------

Co-authored-by: Adrian Catangiu <adrian@parity.io>